NextGP.jl: Modules and Functions

NextGP.jl relies on several other Julia packages, which could be easily avoided and may be removed in the later releases.

Currently one core dependency is the StatsModels.jl package, for model expression, and fixed effect design matrix generation.

Checking StatsModels.jl's manual for at least formula and categorical data could be useful.

Basic Model

NextGP.jl uses the following basic model:
$
\mathbf{y}= \mathbf{X}\mathbf{b} + \sum{i}\mathbf{Z}{i}\mathbf{u}{i} + \sum{j}\mathbf{M}{j}\boldsymbol{\beta}{j} + \mathbf{e} $

\[\mathbf{y}\]

is a vector of phenotypes corrected $\mathbf{X}$ is a matrix of fixed effects $\mathbf{b}$ is a vector of fixed effects $\mathbf{Z}_i$ are matrices of random effects $\mathbf{u}_i$ are vectors of random effects $\mathbf{M}_{j}$ are matrices of random effects $\boldsymbol{\beta}_j$ are vectors of random marker effects $\mathbf{e}$ is the vector of random environmental effects

Public Functions

Public functions of NextGP are documented here

NextGP.prepMatVec.prep — Function

function prep(f::StatsModels.TermOrTerms, inputData::DataFrame;userHints::Dict,path2ped,paths2geno)

NextGP relies on StatsModels.jl package for model expression (f), and fixed effect design matrix generation.

Details for the model expression (f), and fixed effects coding specifications (e.g., effect or dummy coding) can be found at StatsModels.jl.

Design matrices for random effects are generated either own internal functions or using StatsModels.jls modelcols, depending on how user defined the model term in the model.

Reads in marker data, and mean-centers the columns.

Finally returns lhs vector and rhs matrices.

By default:

all Int rhs variables are made Categorical,

all String rhs variables (also those made Categorical) are dummy coded, except those defined by the user in userHints,

all Float rhs variables are centered.

NextGP.makeA — Function

makeA(s::Any, d::Any)

Makes pedigree-based relationship matrix. adapted from http://morotalab.org/Mrode2005/relmat/createA.txt

NextGP.makePed — Function

makePed(inputFile::String,userData)

Makes pedigree using PedigreeBase package

Internals

Functions that are used internally are documented here