About

NextGP relies on several other Julia packages, which could be easily avoided and may be removed in the later releases.

Currently one core dependency is the StatsModels.jl package, for model expreession, and fixed effects design matrices generation.

I suggest you first have a look at least StatsModels.jl and StatsModels.jl

Public Functions

Public functions of NextGP are documented here

NextGP.equations.mme — Function

function mme(f::StatsModels.TermOrTerms, userData::DataFrame;userHints::Dict,blocks,path2ped,paths2geno)

NextGP relies on StatsModels.jl package for model expression (f), and fixed effect design matrix generation.

Details for the model expression (f), and fixed effects coding specifications (e.g., dummy or contrast coding) can be found at StatsModels.jl.

Design matrices for random effects are generated either own internal functions or using StatsModels.jl "modelcols", depending on how user defined the Term in the model.

Reads in marker data, and mean-centers the columns.

Finally returns lhs vector and rhs matrices.

By default:

all Int rhs variables are made Categorical,

all String rhs variables (also those made Categorical) are dummy coded, except those defined by the user in "userHints",

all Float rhs variables are centered.

NextGP.makeA — Function

makeA(s::Any, d::Any)

Makes pedigree-based relationship matrix. adapted from http://morotalab.org/Mrode2005/relmat/createA.txt

NextGP.makePed — Function

makePed(inputFile::String,userData)

Makes pedigree using PedigreeBase package

Internals

Functions that are used internally are documented here